requirements.txt
Flask==3.1.3
gunicorn==25.3.0
pymongo==4.16.0
dnspython==2.8.0
language-tool-python==3.3.0
contractions==0.1.73
textblob==0.20.0
nltk==3.9.4
regex==2026.3.32
requests==2.33.1
tqdm==4.67.3
openai-whisper
ffmpeg-python
imageio-ffmpeg
deepgram-sdk